backend/requirements.txt
aiohttp==3.7.4
aiosignal==1.2.0
altgraph==0.17.2
amqp==5.1.1
anyio==3.6.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.5.0
astroid==2.15.0
asttokens==2.0.5
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
auto-py-to-exe==2.22.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bottle==0.12.23
bottle-websocket==0.2.9
CacheControl==0.12.11
cached-property==1.5.2
cachetools==4.2.2
captcha==0.3
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
cinemagoer==2022.2.11
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudscraper==1.2.60
colorama==0.4.5
colour==0.1.5
cowpy==1.1.5
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.2
django-appconf==1.0.5
django-celery-beat==2.3.0
django-ckeditor==6.2.0
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-heroku==0.3.1
django-htmlmin==0.11.0
django-jalali==6.0.0
django-js-asset==2.0.0
django-recaptcha==2.0.6
django-render-partial==0.4
django-taggit==3.1.0
django-telethon==1.1.3
django-telethon-session==0.9
django-timezone-field==5.0
djangorestframework==3.14.0
Eel==0.14.0
emoji==1.7.0
executing==0.9.1
fake-useragent==0.1.11
filelock==3.0.12
firebase-admin==6.0.0
fonttools==4.34.4
frozenlist==1.3.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.27
glitch-this==1.0.2
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.49.1
grpcio-status==1.49.1
gTTS==2.2.4
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hachoir==3.1.3
heroku3==5.1.4
hpack==4.0.0
html-telegraph-poster==0.4.0
html5lib==1.1
httpcore==0.15.0
httplib2==0.20.4
httptools==0.4.0
httpx==0.23.0
humanize==4.2.3
hyperframe==6.0.1
idna==2.10
IMDbPY==2022.7.9
importlib-metadata==4.12.0
install==1.3.5
ipython==8.4.0
isodate==0.6.0
isort==5.12.0
jdatetime==3.6.2
jedi==0.18.1
jikanpy==4.3.2
JustWatch==0.5.1
kombu==5.2.4
lazy-object-proxy==1.9.0
lottie==0.6.11
lxml==4.6.5
lyricsgenius==3.0.1
Markdown==3.4.1
matplotlib-inline==0.1.3
mccabe==0.7.0
motor==3.0.0
msgpack==1.0.4
multidict==6.0.2
numpy==1.23.1
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
playsound==1.3.0
prettytable==3.3.0
prompt-toolkit==3.0.30
proto-plus==1.22.1
protobuf==4.21.7
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydub==0.25.1
pyfiglet==0.8.post1
pygame==2.1.2
PyGithub==1.55
Pygments==2.12.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyJWT==2.4.0
pylast==5.0.0
pylint==2.17.1
pymediainfo==5.1.0
pymongo==4.2.0
PyMuPDF==1.20.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
Pyrogram==2.0.35
pySmartDL==1.3.4
PySocks==1.7.1
pyTelegramBotAPI==4.4.1
python-barcode==0.14.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-memcached==1.59
python-telegram-bot==13.13
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
qrcode==7.3.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rjsmin==1.1.0
rsa==4.9
schedule==1.1.0
search-engine-parser==0.6.7
selenium==4.3.0
sentry-sdk==1.5.10
ShazamAPI==0.0.2
simplejson==3.17.6
six==1.15.0
smmap==5.0.0
sniffio==1.2.0
sorl-thumbnail==12.6.3
sortedcontainers==2.4.0
soupsieve==2.0.1
spamwatch==0.3.0
speedtest-cli==2.1.3
SQLAlchemy==1.3.23
sqlalchemy-json==0.5.0
sqlparse==0.3.1
stack-data==0.3.0
Telethon==1.24.0
termcolor==1.1.0
termcolor2==0.0.3
TgCrypto==1.2.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.18.2
vine==5.0.0
virtualenv==20.0.31
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
whichcraft==0.6.1
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.1.0
yarl==1.8.1
zeep==3.4.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
